+++ Llama/Llama.Data/Extensions/ILlamaTokenCollectionExtensions.cs
@@ -53,14 +53,14 @@ namespace Llama.Data.Extensions
             return false;
         }
 
-        public static int FindIndex<T>(this IEnumerable<T> source, int start, Func<T, bool> func)
+        public static int FindIndex<T>(this IEnumerable<T> source, uint start, Func<T, bool> func)
         {
-            int i = 0;
+            uint i = 0;
             foreach (T t in source)
             {
                 if (i >= start && func.Invoke(t))
                 {
-                    return i;
+                    return (int)i;
                 }
 
                 i++;
@@ -69,10 +69,10 @@ namespace Llama.Data.Extensions
             return -1;
         }
 
-        public static LlamaTokenCollection From(this ILlamaTokenCollection target, int startIndex, LlamaToken startToken)
+        public static LlamaTokenCollection From(this ILlamaTokenCollection target, uint startIndex, LlamaToken startToken)
         {
             // Calculate the index to start from
-            int start = target.Count - startIndex;
+            uint start = target.Count - startIndex;
 
             // Ensure the index is within valid bounds
             if (start < 0)
@@ -90,7 +90,7 @@ namespace Llama.Data.Extensions
             // If startToken was not found, use the original start position
             if (index == -1)
             {
-                index = start;
+                index = (int)start;
             }
 
             // Copy from the found position (or the original start position if startToken was not found)
@@ -122,7 +122,7 @@ namespace Llama.Data.Extensions
 
                 if (isMatch)
                 {
-                    i += toFind.Count;
+                    i += (int)toFind.Count;
                     foreach (LlamaToken tokenA in toReplace)
                     {
                         toReturn.Append(tokenA);
