+++ Llama/LlamaApi.Shared/Samplers/RepetitionBlockingSampler.cs
@@ -7,22 +7,22 @@ namespace ChieApi.Samplers
 {
     public class RepetitionBlockingSampler : ISimpleSampler
     {
-        private readonly int _max;
+        private readonly uint _max;
 
-        public RepetitionBlockingSampler(int max)
+        public RepetitionBlockingSampler(uint max)
         {
             this._max = max;
         }
 
         public Task SampleNext(InferenceEnumerator enumerator)
         {
-            int ilen = enumerator.Enumerated.Count;
+            uint ilen = enumerator.Enumerated.Count;
 
             if (ilen >= this._max)
             {
-                int len = ilen;
-                int skip = ilen - _max;
-                List<int> ids = enumerator.Enumerated.Skip(skip).Select(s => s.Id).ToList();
+                uint len = ilen;
+                uint skip = ilen - _max;
+                List<int> ids = enumerator.Enumerated.Skip((int)skip).Select(s => s.Id).ToList();
                 List<int> dist = ids.Distinct().ToList();
 
                 if (dist.Count == 1)
